+++ web/src/locales/zh-CN/component.ts
@@ -41,9 +41,9 @@ export default {
   'component.global.edit.plugin': '配置插件',
   'component.global.loading': '加载中',
   'component.global.list': '列表',
-  'component.global.description': '描述',
+  'component.global.description': '描述信息',
   'component.global.labels': '标签',
-  'component.global.version': '版本',
+  'component.global.version': '路由版本',
   'component.global.operation': '操作',
   'component.status.success': '成功',
   'component.status.fail': '失败',
