+++ web/src/pages/Route/components/Step1/RequestConfigView.tsx
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import Form from 'antd/es/form';
-import { Button, Input, Select, Row, Col, InputNumber, Switch } from 'antd';
+import { Button, Input, Select, Row, Col, InputNumber } from 'antd';
 import { PlusOutlined, MinusCircleOutlined } from '@ant-design/icons';
 import { useIntl } from 'umi';
 import { PanelSection } from '@api7-dashboard/ui';
@@ -25,7 +25,6 @@ import {
   HTTP_METHOD_OPTION_LIST,
   FORM_ITEM_WITHOUT_LABEL,
 } from '@/pages/Route/constants';
-import { fetchServiceList } from '../../service';
 
 const removeBtnStyle = {
   marginLeft: 20,
@@ -36,14 +35,8 @@ const removeBtnStyle = {
 const RequestConfigView: React.FC<RouteModule.Step1PassProps> = ({
   form,
   disabled,
-  onChange = () => { },
 }) => {
   const { formatMessage } = useIntl();
-  const [serviceList, setServiceList] = useState<ServiceModule.ResponseBody[]>([]);
-
-  useEffect(() => {
-    fetchServiceList().then(({ data }) => setServiceList(data));
-  }, []);
 
   const HostList = () => (
     <Form.List name="hosts">
@@ -305,143 +298,7 @@ const RequestConfigView: React.FC<RouteModule.Step1PassProps> = ({
     </Form.Item>
   )
 
-  const WebSocket: React.FC = () => (
-    <Form.Item label="WebSocket">
-      <Row>
-        <Col>
-          <Form.Item noStyle valuePropName="checked" name="enable_websocket">
-            <Switch disabled={disabled} />
-          </Form.Item>
-        </Col>
-      </Row>
-    </Form.Item>
-  )
 
-  const Redirect: React.FC = () => {
-    const list = [
-      {
-        value: "forceHttps",
-        label: formatMessage({ id: 'page.route.select.option.enableHttps' })
-      }, {
-        value: "customRedirect",
-        label: formatMessage({ id: 'page.route.select.option.configCustom' })
-      }, {
-        value: "disabled",
-        label: formatMessage({ id: 'page.route.select.option.forbidden' })
-      }
-    ]
-
-    return (
-      <Form.Item label={formatMessage({ id: 'page.route.form.itemLabel.redirect' })}>
-        <Row>
-          <Col span={5}>
-            <Form.Item
-              name="redirectOption"
-              noStyle
-            >
-              <Select
-                disabled={disabled}
-                onChange={(parmas) => {
-                  onChange({ action: 'redirectOptionChange', data: parmas });
-                }}
-              >
-                {list.map(item => (
-                  <Select.Option value={item.value} key={item.value}>
-                    {item.label}
-                  </Select.Option>
-                ))}
-              </Select>
-            </Form.Item>
-          </Col>
-        </Row>
-      </Form.Item>
-    )
-  }
-
-  const CustomRedirect: React.FC = () => (
-    <Form.Item
-      noStyle
-      shouldUpdate={(prev, next) => {
-        if (prev.redirectOption !== next.redirectOption) {
-          onChange({ action: 'redirectOptionChange', data: next.redirectOption });
-        }
-        return prev.redirectOption !== next.redirectOption;
-      }}
-    >
-      {() => {
-        if (form.getFieldValue('redirectOption') === 'customRedirect') {
-          return (
-            <Form.Item
-              label={formatMessage({ id: 'page.route.form.itemLabel.redirectCustom' })}
-              required
-            >
-              <Row gutter={10}>
-                <Col>
-                  <Form.Item
-                    name="redirectURI"
-                    rules={[
-                      {
-                        required: true,
-                        message: `${formatMessage({
-                          id: 'component.global.pleaseEnter',
-                        })}${formatMessage({
-                          id: 'page.route.form.itemLabel.redirectURI',
-                        })}`,
-                      },
-                    ]}
-                  >
-                    <Input
-                      placeholder={formatMessage({
-                        id: 'page.route.input.placeholder.redirectCustom',
-                      })}
-                      disabled={disabled}
-                    />
-                  </Form.Item>
-                </Col>
-                <Col span={10}>
-                  <Form.Item name="ret_code" rules={[{ required: true }]}>
-                    <Select disabled={disabled}>
-                      <Select.Option value={301}>
-                        {formatMessage({ id: 'page.route.select.option.redirect301' })}
-                      </Select.Option>
-                      <Select.Option value={302}>
-                        {formatMessage({ id: 'page.route.select.option.redirect302' })}
-                      </Select.Option>
-                    </Select>
-                  </Form.Item>
-                </Col>
-              </Row>
-            </Form.Item>
-          );
-        }
-        return null;
-      }}
-    </Form.Item>
-  )
-
-  const ServiceSelector: React.FC = () => (
-    <Form.Item label={formatMessage({ id: 'page.route.service' })}>
-      <Row>
-        <Col span={5}>
-          <Form.Item noStyle name="service_id">
-            <Select disabled={disabled}>
-              {/* TODO: value === '' means  no service_id select, need to find a better way */}
-              <Select.Option value="" key={Math.random().toString(36).substring(7)}>
-                {formatMessage({ id: "page.route.service.none" })}
-              </Select.Option>
-              {serviceList.map((item) => {
-                return (
-                  <Select.Option value={item.id} key={item.id}>
-                    {item.name}
-                  </Select.Option>
-                );
-              })}
-            </Select>
-          </Form.Item>
-        </Col>
-      </Row>
-    </Form.Item>
-  )
 
   return (
     <PanelSection
@@ -452,12 +309,6 @@ const RequestConfigView: React.FC<RouteModule.Step1PassProps> = ({
       <RemoteAddrList />
       <HTTPMethods />
       <RoutePriority />
-      <WebSocket />
-
-      <Redirect />
-      <CustomRedirect />
-
-      <ServiceSelector />
     </PanelSection>
   );
 };
