+++ web/src/pages/Route/locales/zh-CN.ts
@@ -36,12 +36,12 @@ export default {
   'page.route.status': '状态',
   'page.route.groupName': '分组名称',
   'page.route.offline': '下线',
-  'page.route.publish': '发布',
+  'page.route.publish': '是否发布',
   'page.route.published': '已发布',
   'page.route.unpublished': '未发布',
   'page.route.onlineDebug': '在线调试',
   'page.route.pluginTemplateConfig': '插件模版配置',
-  'page.route.service': '服务',
+  'page.route.service': '绑定服务',
   'page.route.instructions': '说明',
   'page.route.import': '导入',
   'page.route.createRoute': '创建路由',
